refactor(data): split local data sources by feature and add typed list query router
