--- app/src/main/java/com/example/lifetogether/data/local/LocalDataSource.kt
@@ -1,1392 +1,0 @@
-package com.example.lifetogether.data.local
-
-import android.content.ContentValues
-import android.content.Context
-import android.net.Uri
-import android.provider.MediaStore
-import android.util.Log
-import androidx.core.net.toUri
-import com.example.lifetogether.data.local.dao.AlbumsDao
-import com.example.lifetogether.data.local.dao.CategoriesDao
-import com.example.lifetogether.data.local.dao.FamilyInformationDao
-import com.example.lifetogether.data.local.dao.GalleryMediaDao
-import com.example.lifetogether.data.local.dao.GuideProgressDao
-import com.example.lifetogether.data.local.dao.GuidesDao
-import com.example.lifetogether.data.local.dao.GroceryListDao
-import com.example.lifetogether.data.local.dao.GrocerySuggestionsDao
-import com.example.lifetogether.data.local.dao.RoutineListsDao
-import com.example.lifetogether.data.local.dao.UserListsDao
-import com.example.lifetogether.data.local.dao.RecipesDao
-import com.example.lifetogether.data.local.dao.TipTrackerDao
-import com.example.lifetogether.data.local.dao.UserInformationDao
-import com.example.lifetogether.data.logic.generateImageThumbnailFromFile
-import com.example.lifetogether.data.logic.generateVideoThumbnailFromFile
-import com.example.lifetogether.data.model.AlbumEntity
-import com.example.lifetogether.data.model.CategoryEntity
-import com.example.lifetogether.data.model.Entity
-import com.example.lifetogether.data.model.FamilyEntity
-import com.example.lifetogether.data.model.FamilyMemberEntity
-import com.example.lifetogether.data.model.GalleryMediaEntity
-import com.example.lifetogether.data.model.GuideProgressEntity
-import com.example.lifetogether.data.model.GuideEntity
-import com.example.lifetogether.data.model.GroceryListEntity
-import com.example.lifetogether.data.model.GrocerySuggestionEntity
-import com.example.lifetogether.data.model.RoutineListEntryEntity
-import com.example.lifetogether.data.model.UserListEntity
-import com.example.lifetogether.data.model.RecipeEntity
-import com.example.lifetogether.data.model.TipEntity
-import com.example.lifetogether.data.model.UserEntity
-import com.example.lifetogether.domain.listener.ResultListener
-import com.example.lifetogether.domain.logic.GuideProgressSnapshot
-import com.example.lifetogether.domain.model.Category
-import com.example.lifetogether.domain.model.TipItem
-import com.example.lifetogether.domain.model.UserInformation
-import com.example.lifetogether.domain.model.enums.MediaType
-import com.example.lifetogether.domain.model.family.FamilyInformation
-import com.example.lifetogether.domain.model.gallery.Album
-import com.example.lifetogether.domain.model.gallery.GalleryImage
-import com.example.lifetogether.domain.model.gallery.GalleryMedia
-import com.example.lifetogether.domain.model.gallery.GalleryVideo
-import com.example.lifetogether.domain.model.guides.Guide
-import com.example.lifetogether.domain.model.guides.GuideProgressState
-import com.example.lifetogether.domain.model.grocery.GroceryItem
-import com.example.lifetogether.domain.model.grocery.GrocerySuggestion
-import com.example.lifetogether.domain.model.lists.RoutineListEntry
-import com.example.lifetogether.domain.model.lists.UserList
-import com.example.lifetogether.domain.model.recipe.Recipe
-import com.example.lifetogether.domain.model.sealed.ImageType
-import com.example.lifetogether.util.Constants
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.mapNotNull
-import kotlinx.coroutines.withContext
-import java.io.File
-import java.util.Date
-import javax.inject.Inject
-
-class LocalDataSource @Inject constructor(
-    @ApplicationContext private val context: Context, // Injected app context
-    private val groceryListDao: GroceryListDao,
-    private val recipesDao: RecipesDao,
-    private val grocerySuggestionsDao: GrocerySuggestionsDao,
-    private val categoriesDao: CategoriesDao,
-    private val userInformationDao: UserInformationDao,
-    private val familyInformationDao: FamilyInformationDao,
-    private val galleryMediaDao: GalleryMediaDao,
-    private val albumsDao: AlbumsDao,
-    private val tipTrackerDao: TipTrackerDao,
-    private val guidesDao: GuidesDao,
-    private val guideProgressDao: GuideProgressDao,
-    private val userListsDao: UserListsDao,
-    private val routineListsDao: RoutineListsDao,
-) {
-    companion object {
-        private const val TAG = "LocalDataSource"
-    }
-    // -------------------------------------------------------------- CATEGORIES
-    fun getCategories(): Flow<List<CategoryEntity>> {
-        return categoriesDao.getItems()
-    }
-    suspend fun updateCategories(items: List<Category>) {
-        val categoryEntities = items.map { category ->
-            CategoryEntity(
-                emoji = category.emoji,
-                name = category.name,
-            )
-        }
-        categoriesDao.updateItems(categoryEntities)
-
-        // Fetch the current items from the Room database
-        val currentItems = categoriesDao.getItems().first()
-
-        val itemsToUpdate = categoryEntities.filter { newItem ->
-            currentItems.none { currentItem -> newItem.name == currentItem.name && newItem.emoji == currentItem.emoji }
-        }
-        val itemsToDelete = currentItems.filter { currentItem ->
-            categoryEntities.none { newItem -> newItem.name == currentItem.name }
-        }
-        categoriesDao.updateItems(itemsToUpdate)
-        categoriesDao.deleteItems(itemsToDelete)
-    }
-
-    // -------------------------------------------------------------- ITEMS
-    fun getListItems(
-        listName: String,
-        familyId: String,
-        uid: String? = null,
-    ): Flow<List<Entity>> {
-        Log.d(TAG, "getListItems listName: $listName")
-        val items: Flow<List<Entity>> = when (listName) {
-            Constants.GROCERY_TABLE -> groceryListDao.getItems(familyId).map { list ->
-                list.map { Entity.GroceryList(it) }
-            }
-            Constants.RECIPES_TABLE -> recipesDao.getItems(familyId).map { list ->
-                list.map { Entity.Recipe(it) }
-            }
-            Constants.ALBUMS_TABLE -> albumsDao.getItems(familyId).map { list ->
-                list.map { Entity.Album(it) }
-            }
-            Constants.GALLERY_MEDIA_TABLE -> galleryMediaDao.getItems(familyId).map { list ->
-                list.map { Entity.GalleryMedia(it) }
-            }
-            Constants.TIP_TRACKER_TABLE -> tipTrackerDao.getItems(familyId).map { list ->
-                list.map { Entity.Tip(it) }
-            }
-            Constants.GUIDES_TABLE -> {
-                if (uid.isNullOrBlank()) {
-                    guidesDao.getItems(familyId).map { list ->
-                        list.map { Entity.Guide(it) }
-                    }
-                } else {
-                    combine(
-                        guidesDao.getItems(familyId),
-                        guideProgressDao.getItems(familyId, uid),
-                    ) { guides, progressItems ->
-                        val progressByGuideId = progressItems.associateBy { it.guideId }
-                        guides.map { guide ->
-                            Entity.Guide(
-                                applyProgressToGuideEntity(
-                                    guide = guide,
-                                    progress = progressByGuideId[guide.id],
-                                ),
-                            )
-                        }
-                    }
-                }
-            }
-            Constants.USER_LISTS_TABLE -> userListsDao.getItems(familyId).map { list ->
-                list.map { Entity.UserList(it) }
-            }
-            Constants.ROUTINE_LIST_ENTRIES_TABLE -> {
-                if (uid.isNullOrBlank()) {
-                    routineListsDao.getItems(familyId).map { list -> list.map { Entity.RoutineListEntry(it) } }
-                } else {
-                    routineListsDao.getItemsByListId(familyId, uid).map { list -> list.map { Entity.RoutineListEntry(it) } }
-                }
-            }
-            else -> flowOf(emptyList()) // Handle the case where the listName doesn't match any known entity
-        }
-        return items
-    }
-
-    fun getItemById(
-        listName: String,
-        familyId: String,
-        id: String,
-        uid: String? = null,
-    ): Flow<Entity> {
-        Log.d(TAG, "getItemById listName=$listName familyId=$familyId id=$id")
-        return when (listName) {
-            Constants.RECIPES_TABLE -> flow {
-                val recipe = recipesDao.getItemById(familyId, id)
-                if (recipe != null) {
-                    emit(Entity.Recipe(recipe))
-                }
-            }
-            Constants.ALBUMS_TABLE -> flow {
-                val album = albumsDao.getItemById(familyId, id)
-                if (album != null) {
-                    emit(Entity.Album(album))
-                }
-            }
-            Constants.GALLERY_MEDIA_TABLE -> flow {
-                val galleryMedia = galleryMediaDao.getItemById(familyId, id)
-                if (galleryMedia != null) {
-                    emit(Entity.GalleryMedia(galleryMedia))
-                }
-            }
-            Constants.GUIDES_TABLE -> flow {
-                Log.d(TAG, "getItemById guide flow init familyId=$familyId id=$id")
-                val guideFlow = guidesDao.getItemByIdFlow(familyId, id)
-                val mergedFlow = if (uid.isNullOrBlank()) {
-                    guideFlow.map { guide ->
-                        guide?.let {
-                            Entity.Guide(it)
-                        }
-                    }
-                } else {
-                    combine(
-                        guideFlow,
-                        guideProgressDao.getItemByGuideIdFlow(familyId, uid, id),
-                    ) { guide, progress ->
-                        guide?.let {
-                            Entity.Guide(applyProgressToGuideEntity(it, progress))
-                        }
-                    }
-                }
-                emitAll(
-                    mergedFlow.mapNotNull { guideEntity ->
-                        if (guideEntity == null) {
-                            Log.w(TAG, "getItemById guide emission=null familyId=$familyId id=$id")
-                            null
-                        } else {
-                            val guide = guideEntity.entity
-                            Log.d(
-                                TAG,
-                                "getItemById guide emission id=${guide.id} started=${guide.started} resume=${guide.resume} lastUpdated=${guide.lastUpdated.time}",
-                            )
-                            guideEntity
-                        }
-                    },
-                )
-            }
-
-            Constants.ROUTINE_LIST_ENTRIES_TABLE -> flow {
-                val entry = routineListsDao.getItemById(id)
-                if (entry != null) {
-                    emit(Entity.RoutineListEntry(entry))
-                }
-            }
-            else -> flowOf() // Handle the case where the listName doesn't match any known entity
-        }
-    }
-
-    fun deleteItems(
-        listName: String,
-        itemIds: List<String>,
-    ): ResultListener {
-        Log.d(TAG, "deleteItems()")
-        try {
-            when (listName) {
-                Constants.GROCERY_TABLE -> groceryListDao.deleteItems(itemIds)
-                else -> {}
-            }
-            return ResultListener.Success
-        } catch (e: Exception) {
-            return ResultListener.Failure("Error: ${e.message}")
-        }
-    }
-
-    // -------------------------------------------------------------- RECIPES
-    suspend fun updateRecipes(
-        items: List<Recipe>,
-        byteArrays: Map<String, ByteArray>,
-    ) {
-        Log.d(TAG, "updateRecipes()")
-
-        var recipeEntityList = items.map { item ->
-            RecipeEntity(
-                id = item.id ?: "",
-                familyId = item.familyId,
-                itemName = item.itemName,
-                lastUpdated = item.lastUpdated,
-                description = item.description,
-                ingredients = item.ingredients,
-                instructions = item.instructions,
-                preparationTimeMin = item.preparationTimeMin,
-                favourite = item.favourite,
-                servings = item.servings,
-                tags = item.tags,
-            )
-        }
-
-        if (byteArrays.isNotEmpty()) {
-            recipeEntityList = recipeEntityList.map { item ->
-                item.copy(imageData = if (byteArrays[item.id] != null) byteArrays[item.id] else null)
-            }
-        }
-
-        // Fetch the current items from the Room database
-        val currentItems = recipesDao.getItems(items[0].familyId).first()
-
-        // Determine the items to be inserted or updated
-        val itemsToUpdate = recipeEntityList.filter { newItem ->
-            currentItems.none { currentItem -> newItem.id == currentItem.id && newItem == currentItem }
-        }
-
-        // Determine the items to be deleted
-        val itemsToDelete = currentItems.filter { currentItem ->
-            recipeEntityList.none { newItem -> newItem.id == currentItem.id }
-        }
-
-        // Update the Room database with the new or changed items
-        recipesDao.updateItems(itemsToUpdate)
-
-        // Delete the items that no longer exist in Firestore
-        recipesDao.deleteItems(itemsToDelete.map { it.id })
-    }
-
-    suspend fun deleteFamilyRecipes(familyId: String) {
-        val currentFamilyItems = recipesDao.getItems(familyId).firstOrNull()
-        if (currentFamilyItems != null) {
-            recipesDao.deleteItems(currentFamilyItems.map { it.id })
-        }
-    }
-
-    suspend fun getRecipeIdsWithImages(familyId: String): Set<String> {
-        return recipesDao.getRecipeIdsWithImages(familyId).toSet()
-    }
-
-    // -------------------------------------------------------------- GROCERY LIST
-    suspend fun updateGroceryList(items: List<GroceryItem>) {
-        Log.d(TAG, "updateGroceryList()")
-
-        val groceryListEntityList = items.map { item ->
-            GroceryListEntity(
-                id = item.id ?: "",
-                familyId = item.familyId,
-                name = item.itemName,
-                lastUpdated = item.lastUpdated,
-                completed = item.completed,
-                category = item.category,
-                approxPrice = item.approxPrice,
-            )
-        }
-
-        // Fetch the current items from the Room database
-        val currentItems = groceryListDao.getItems(items[0].familyId).first()
-
-        // Determine the items to be inserted or updated
-        val itemsToUpdate = groceryListEntityList.filter { newItem ->
-            currentItems.none { currentItem -> newItem.id == currentItem.id && newItem == currentItem }
-        }
-
-        // Determine the items to be deleted
-        val itemsToDelete = currentItems.filter { currentItem ->
-            groceryListEntityList.none { newItem -> newItem.id == currentItem.id }
-        }
-
-        // Update the Room database with the new or changed items
-        groceryListDao.updateItems(itemsToUpdate)
-
-        // Delete the items that no longer exist in Firestore
-        groceryListDao.deleteItems(itemsToDelete.map { it.id })
-    }
-
-    suspend fun deleteFamilyGroceryItems(familyId: String) {
-        val currentFamilyItems = groceryListDao.getItems(familyId).firstOrNull()
-        if (currentFamilyItems != null) {
-            groceryListDao.deleteItems(currentFamilyItems.map { it.id })
-        }
-    }
-
-    // -------------------------------------------------------------- GROCERY SUGGESTIONS
-    fun getGrocerySuggestions(): Flow<List<GrocerySuggestionEntity>> {
-        return grocerySuggestionsDao.getItems()
-    }
-
-    suspend fun updateGrocerySuggestions(items: List<GrocerySuggestion>) {
-        Log.d(TAG, "updateGrocerySuggestions()")
-        val grocerySuggestionEntities = items.mapNotNull { grocerySuggestion ->
-            grocerySuggestion.id?.let { id ->
-                GrocerySuggestionEntity(
-                    id = id,
-                    suggestionName = grocerySuggestion.suggestionName,
-                    category = grocerySuggestion.category,
-                    approxPrice = grocerySuggestion.approxPrice,
-                )
-            }
-        }
-        grocerySuggestionsDao.updateItems(grocerySuggestionEntities)
-
-        // Fetch the current items from the Room database
-        //   getItems() returns a Flow, so you need to use first() to get the current value
-        val currentItems = grocerySuggestionsDao.getItems().first()
-
-        // Determine the items to be inserted or updated
-        val itemsToUpdate = grocerySuggestionEntities.filter { newItem ->
-            currentItems.none { currentItem -> newItem.id == currentItem.id }
-        }
-
-        // Determine the items to be deleted
-        val itemsToDelete = currentItems.filter { currentItem ->
-            grocerySuggestionEntities.none { newItem -> newItem.id == currentItem.id }
-        }
-
-        // Update the Room database with the new or changed items
-        grocerySuggestionsDao.updateItems(itemsToUpdate)
-
-        // Delete the items that no longer exist in Firestore
-        grocerySuggestionsDao.deleteItems(itemsToDelete)
-    }
-
-    // -------------------------------------------------------------- USER INFORMATION
-    fun getUserInformation(uid: String): Flow<UserEntity?> {
-        return userInformationDao.getItems(uid)
-    }
-
-    suspend fun updateUserInformation(
-        userInformation: UserInformation,
-        byteArray: ByteArray? = null,
-    ) {
-        var userEntity = UserEntity(
-            uid = userInformation.uid ?: "",
-            email = userInformation.email,
-            name = userInformation.name,
-            birthday = userInformation.birthday,
-            familyId = userInformation.familyId,
-        )
-        if (byteArray != null) {
-            userEntity = userEntity.copy(imageData = byteArray)
-        }
-
-        userInformationDao.updateItems(userEntity)
-    }
-
-    suspend fun userHasProfileImage(uid: String): Boolean {
-        return (userInformationDao.hasImageData(uid) ?: 0) == 1
-    }
-
-    fun clearUserInformationTables(): ResultListener {
-        try {
-            groceryListDao.deleteTable()
-            recipesDao.deleteTable()
-            userInformationDao.deleteTable()
-            familyInformationDao.deleteFamiliesTable()
-            familyInformationDao.deleteFamilyMembersTable()
-
-            val resolver = context.contentResolver
-            val galleryImages = galleryMediaDao.getAll()
-            galleryImages.forEach { item ->
-                item.mediaUri?.let { resolver.delete(it.toUri(), null, null) }
-            }
-
-            galleryMediaDao.deleteTable()
-            albumsDao.deleteTable()
-            guidesDao.deleteTable()
-
-            return ResultListener.Success
-        } catch (e: Exception) {
-            return ResultListener.Failure("Error: $e")
-        }
-    }
-
-    // -------------------------------------------------------------- FAMILY INFORMATION
-    fun getFamilyInformation(familyId: String): Flow<FamilyEntity> {
-        return familyInformationDao.getFamilyInfo(familyId)
-    }
-
-    fun getFamilyMembers(familyId: String): Flow<List<FamilyMemberEntity>> {
-        return familyInformationDao.getFamilyMembers(familyId)
-    }
-
-    suspend fun updateFamilyInformation(
-        familyInformation: FamilyInformation,
-        byteArray: ByteArray? = null,
-    ) {
-        // Update FamilyEntity
-        var familyEntity = FamilyEntity(
-            familyId = familyInformation.familyId ?: "",
-        )
-
-        if (byteArray != null) {
-            familyEntity = familyEntity.copy(imageData = byteArray)
-        }
-
-        // Update FamilyEntity in the database
-        familyInformationDao.updateFamily(familyEntity)
-
-        // Update Family Members
-        val familyMembers = familyInformation.members?.map {
-            FamilyMemberEntity(
-                uid = it.uid ?: "",
-                familyId = familyInformation.familyId,
-                name = it.name,
-            )
-        } ?: emptyList()
-
-        // Insert or update FamilyMemberEntity
-        familyInformationDao.updateFamilyMembers(familyMembers)
-    }
-
-    suspend fun familyHasImage(familyId: String): Boolean {
-        return (familyInformationDao.hasImageData(familyId) ?: 0) == 1
-    }
-
-    // -------------------------------------------------------------- GALLERY MEDIA
-    fun getAlbumMedia(
-        familyId: String,
-        albumId: String,
-    ): Flow<List<Entity.GalleryMedia>> {
-        Log.d(TAG, "getAlbumMedia")
-        val items: Flow<List<Entity.GalleryMedia>> =
-            galleryMediaDao.getItemsByAlbumId(familyId, albumId)
-                .map { list ->
-                    list.map { Entity.GalleryMedia(it) }
-                }
-        return items
-    }
-
-    suspend fun getAlbumMediaThumbnail(
-        imageId: String,
-    ): ByteArray? {
-        // If thumbnail is missing, try to regenerate it from the stored mediaUri
-        val existing = galleryMediaDao.getMediaThumbnail(imageId)
-        if (existing != null) return existing
-
-        return regenerateAndPersistThumbnail(imageId)
-    }
-
-    suspend fun getAlbumThumbnail(
-        albumId: String,
-    ): ByteArray? {
-        val existing = galleryMediaDao.getNewestMediaThumbnailByAlbumId(albumId)
-        if (existing != null) return existing
-
-        // Fallback: find the latest media in the album and regenerate its thumbnail
-        val latestMediaId = galleryMediaDao.getLatestMediaIdForAlbum(albumId)
-        return latestMediaId?.let { regenerateAndPersistThumbnail(it) }
-    }
-
-    private suspend fun regenerateAndPersistThumbnail(mediaId: String): ByteArray? {
-        val entity = galleryMediaDao.getItemByIdDirect(mediaId) ?: return null
-        val mediaUri = entity.mediaUri?.toUri() ?: return null
-
-        // Copy content to a temp file to reuse existing thumbnail generators
-        val tempFile = withContext(Dispatchers.IO) {
-            File.createTempFile("thumb_regen_", "tmp", context.cacheDir)
-        }
-        try {
-            context.contentResolver.openInputStream(mediaUri)?.use { input ->
-                tempFile.outputStream().use { output -> input.copyTo(output) }
-            } ?: return null
-
-            val regenerated = when (entity.mediaType) {
-                MediaType.IMAGE -> generateImageThumbnailFromFile(
-                    tempFile,
-                )
-
-                MediaType.VIDEO -> generateVideoThumbnailFromFile(
-                    tempFile,
-                )
-            }
-
-            if (regenerated != null) {
-                // Persist the regenerated thumbnail so we don't do this again
-                galleryMediaDao.updateItems(listOf(entity.copy(thumbnail = regenerated)))
-            }
-            return regenerated
-        } catch (e: Exception) {
-            Log.e("LocalDataSource", "Failed to regenerate thumbnail for $mediaId: ${e.message}", e)
-            return null
-        } finally {
-            if (tempFile.exists()) tempFile.delete()
-        }
-    }
-
-    suspend fun getExistingGalleryMediaInfo(
-        familyId: String,
-    ): Map<String, Pair<String?, String?>> {
-        // Returns a map of mediaId to (mediaUri, mediaUrl)
-        // Since we don't store mediaUrl, we return null for it
-        // This checks if media exists locally by checking if mediaUri exists
-        val entities = galleryMediaDao.getExistingMediaIdsWithUris(familyId)
-        return entities.associate { entity ->
-            entity.id to Pair(entity.mediaUri, null)
-        }
-    }
-
-    suspend fun updateGalleryMedia(
-        familyId: String,
-        items: List<Pair<GalleryMedia, File>>,
-        completeSourceList: List<GalleryMedia>? = null, // Complete list from Firestore to detect deletions
-    ) {
-        // Allow empty items list if completeSourceList is provided (for deletion-only operations)
-        if (items.isEmpty() && completeSourceList == null) {
-            return
-        }
-        Log.d(TAG, "updateGalleryMedia(): Saving images and videos to MediaStore")
-
-        // Fetch current items from Room
-        val currentRoomItems = galleryMediaDao.getItems(familyId).firstOrNull() ?: emptyList()
-        val currentMediaStoreUrisMap = currentRoomItems.associateBy({ it.id }, { it.mediaUri })
-
-        val entityList = mutableListOf<GalleryMediaEntity>()
-
-        // If we already have all files locally and only metadata changed (e.g., album move),
-        // rebuild entities from the complete Firestore list so we can upsert metadata changes
-        // without re-downloading the media files.
-        if (items.isEmpty() && completeSourceList != null) {
-            val currentItemsById = currentRoomItems.associateBy { it.id }
-            completeSourceList.forEach { mediaItem ->
-                val existing = mediaItem.id?.let { currentItemsById[it] } ?: return@forEach
-                entityList.add(
-                    existing.copy(
-                        albumId = mediaItem.albumId,
-                        itemName = mediaItem.itemName,
-                        lastUpdated = mediaItem.lastUpdated,
-                    ),
-                )
-            }
-        }
-
-        for ((mediaItem, downloadedFile) in items) {
-            Log.d(
-                "LocalDataSource",
-                "Processing item: ${mediaItem.itemName} (ID: ${mediaItem.id}), File: ${downloadedFile.name}",
-            )
-
-            // 1. Save to internal storage if new, or keep existing URI
-            val mediaStoragePathString: String? = currentMediaStoreUrisMap[mediaItem.id]
-                ?: saveMediaFileToInternalStorage(context, downloadedFile, mediaItem)
-
-            Log.d(
-                "LocalDataSource",
-                "Media URI for ${mediaItem.itemName} (ID: ${mediaItem.id}): $mediaStoragePathString",
-            )
-
-            if (mediaStoragePathString == null && currentMediaStoreUrisMap[mediaItem.id] == null) {
-                Log.w(
-                    "LocalDataSource",
-                    "Failed to save ${mediaItem.itemName} to internal storage and no prior URI existed. Skipping.",
-                )
-                // downloadedFile.delete() // Consider deleting temp file if it's truly temporary and save failed
-                continue
-            }
-
-            // 2. Generate thumbnail from the file
-            // Pass mediaItem to help generateThumbnailBytesFromFile determine type if needed,
-            // though it can also infer from file extension or content.
-            val thumbnailBytes = when (mediaItem) {
-                is GalleryImage -> generateImageThumbnailFromFile(downloadedFile)
-                is GalleryVideo -> generateVideoThumbnailFromFile(downloadedFile)
-            }
-
-            if (thumbnailBytes == null) {
-                Log.w(
-                    "LocalDataSource",
-                    "Failed to generate thumbnail for ${mediaItem.itemName}. It might be stored without a thumbnail.",
-                )
-                // Decide if you want to proceed without a thumbnail or skip
-            }
-
-            val videoDuration = if (mediaItem is GalleryVideo) {
-                mediaItem.duration
-            } else {
-                null
-            }
-
-            entityList.add(
-                GalleryMediaEntity(
-                    id = mediaItem.id!!,
-                    mediaType = mediaItem.mediaType,
-                    familyId = mediaItem.familyId,
-                    itemName = mediaItem.itemName,
-                    lastUpdated = mediaItem.lastUpdated,
-                    albumId = mediaItem.albumId,
-                    dateCreated = mediaItem.dateCreated,
-                    mediaUri = mediaStoragePathString,
-                    thumbnail = thumbnailBytes,
-                    videoDuration = videoDuration,
-                ),
-            )
-
-            Log.d(
-                "LocalDataSource",
-                "Inserted entity for ${mediaItem.itemName}: mediaUri=$mediaStoragePathString",
-            )
-
-            // 3. IMPORTANT: Delete the temporary downloaded file after processing
-            // Ensure this file is always a temporary one that should be cleaned up.
-            if (downloadedFile.exists()) {
-                val deleted = downloadedFile.delete()
-                Log.d("LocalDataSource", "Temporary file ${downloadedFile.name} deleted: $deleted")
-            }
-        }
-        Log.i(
-            "LocalDataSource",
-            "Finished processing ${entityList.size} items into entities for Room.",
-        )
-
-        // --- Logic for determining Room updates and deletes ---
-        // IMPORTANT: Only delete if we have the complete source list from Firestore
-        // If completeSourceList is null, this is a partial batch and we skip deletion
-
-        val itemsToUpdateOrInsert = entityList.filter { newEntity ->
-            currentRoomItems.none { currentEntity ->
-                newEntity.id == currentEntity.id &&
-                    newEntity.mediaUri == currentEntity.mediaUri && // Key comparison points
-                    newEntity.itemName == currentEntity.itemName &&
-                    newEntity.albumId == currentEntity.albumId &&
-                    newEntity.lastUpdated == currentEntity.lastUpdated &&
-                    (
-                        newEntity.thumbnail?.contentEquals(
-                            currentEntity.thumbnail ?: byteArrayOf(),
-                        ) ?: (currentEntity.thumbnail == null)
-                        )
-                // Add other relevant fields if needed for "sameness"
-            }
-        }
-
-        // Handle deletions only if we have the complete source list
-        if (completeSourceList != null) {
-            val currentIdsInRoom = currentRoomItems.map { it.id }.toSet()
-            val sourceIds = completeSourceList.mapNotNull { it.id }.toSet()
-            val idsToDeleteFromRoom = currentIdsInRoom - sourceIds
-            val itemsToDeleteFromMediaStoreAndRoom =
-                currentRoomItems.filter { it.id in idsToDeleteFromRoom }
-
-            // Delete from MediaStore for items that are no longer in the source list
-            if (itemsToDeleteFromMediaStoreAndRoom.isNotEmpty()) {
-                Log.i(
-                    "LocalDataSource",
-                    "Deleting ${itemsToDeleteFromMediaStoreAndRoom.size} items from MediaStore (not in Firestore).",
-                )
-                itemsToDeleteFromMediaStoreAndRoom.forEach { itemToDelete ->
-                    itemToDelete.mediaUri?.let { uriString ->
-                        try {
-                            // Internal storage files - delete directly
-                            val file = File(uriString)
-                            if (file.exists()) {
-                                val deleted = file.delete()
-                                Log.d(
-                                    "LocalDataSource",
-                                    "Deleted file $uriString: $deleted",
-                                )
-                            }
-                        } catch (e: Exception) {
-                            Log.e(
-                                "LocalDataSource",
-                                "Error deleting file $uriString: ${e.message}",
-                            )
-                        }
-                    }
-                }
-            }
-
-            // Delete from Room for items no longer in the source list
-            if (idsToDeleteFromRoom.isNotEmpty()) {
-                galleryMediaDao.deleteItems(idsToDeleteFromRoom.toList())
-                Log.i("LocalDataSource", "Deleted ${idsToDeleteFromRoom.size} items from Room (not in Firestore).")
-            }
-        } else {
-            Log.d(
-                "LocalDataSource",
-                "Skipping deletion logic - this is a partial sync batch, not a complete source list. Items not in this batch may retry on next sync.",
-            )
-        }
-
-        // Update Room: Upsert new/changed items
-        if (itemsToUpdateOrInsert.isNotEmpty()) {
-            Log.d(
-                "LocalDataSource",
-                "Upserting ${itemsToUpdateOrInsert.size} items: ${itemsToUpdateOrInsert.map { it.itemName + " (${it.mediaUri})" }}",
-            )
-            galleryMediaDao.updateItems(itemsToUpdateOrInsert) // Assuming this is an upsert (inserts or updates)
-            Log.i("LocalDataSource", "Upserted ${itemsToUpdateOrInsert.size} items in Room.")
-        }
-
-        Log.i("LocalDataSource", "updateGalleryMedia finished.")
-    }
-
-    // --- Unified function to save ANY media File (image or video) to Internal Storage ---
-    private fun saveMediaFileToInternalStorage(
-        context: Context,
-        mediaFile: File,
-        mediaItem: GalleryMedia,
-    ): String? {
-        return try {
-            // Create directory for internal storage
-            val mediaDir =
-                File(context.filesDir, "media/${mediaItem.familyId}/${mediaItem.albumId}")
-            if (!mediaDir.exists()) {
-                mediaDir.mkdirs()
-            }
-
-            // Create destination file with original file extension
-            val fileName = "${mediaItem.id}_${mediaItem.itemName}"
-            val destinationFile = File(mediaDir, fileName)
-
-            // Copy file to internal storage
-            mediaFile.inputStream().use { input ->
-                destinationFile.outputStream().use { output ->
-                    input.copyTo(output)
-                }
-            }
-
-            Log.d(
-                "LocalDataSource",
-                "Saved ${mediaItem.itemName} to internal storage: ${destinationFile.absolutePath}",
-            )
-            destinationFile.absolutePath
-        } catch (e: Exception) {
-            Log.e(
-                "LocalDataSource",
-                "Error saving ${mediaItem.itemName} to internal storage: ${e.message}",
-                e,
-            )
-            null
-        }
-    }
-
-    suspend fun deleteFamilyGalleryMedia(familyId: String) {
-        val currentFamilyItems = galleryMediaDao.getItems(familyId).firstOrNull()
-
-        if (currentFamilyItems != null) {
-            galleryMediaDao.deleteItems(currentFamilyItems.map { it.id })
-
-//            val resolver = context.contentResolver
-//
-//            // Delete images from MediaStore
-//            currentFamilyItems.forEach { item ->
-//                item.mediaUri?.let { resolver.delete(it.toUri(), null, null) }
-//            }
-        }
-    }
-
-    // -------------------------------------------------------------- ALBUMS
-    suspend fun updateAlbums(
-        items: List<Album>,
-    ) {
-        Log.d(TAG, "updateAlbums()")
-        val entityList = items.map { item ->
-            AlbumEntity(
-                id = item.id ?: "",
-                familyId = item.familyId,
-                itemName = item.itemName,
-                lastUpdated = item.lastUpdated,
-                count = item.count,
-            )
-        }
-
-        // Fetch the current items from the Room database
-        val currentItems = albumsDao.getItems(items[0].familyId).first()
-
-        // Determine the items to be inserted or updated
-        val itemsToUpdate = entityList.filter { newItem ->
-            currentItems.none { currentItem -> newItem.id == currentItem.id && newItem == currentItem }
-        }
-
-        // Determine the items to be deleted
-        val itemsToDelete = currentItems.filter { currentItem ->
-            entityList.none { newItem -> newItem.id == currentItem.id }
-        }
-
-        // Update the Room database with the new or changed items
-        albumsDao.updateItems(itemsToUpdate)
-
-        // Delete the items that no longer exist in Firestore
-        albumsDao.deleteItems(itemsToDelete.map { it.id })
-    }
-
-    suspend fun deleteFamilyAlbums(familyId: String) {
-        val currentFamilyItems = albumsDao.getItems(familyId).firstOrNull()
-        if (currentFamilyItems != null) {
-            albumsDao.deleteItems(currentFamilyItems.map { it.id })
-        }
-    }
-
-    // -------------------------------------------------------------- GROCERY LIST
-    suspend fun updateTipTracker(items: List<TipItem>) {
-        Log.d(TAG, "updateTipTracker()")
-        val tipEntityList = items.map { item ->
-            TipEntity(
-                id = item.id ?: "",
-                familyId = item.familyId,
-                itemName = item.itemName,
-                lastUpdated = item.lastUpdated,
-                amount = item.amount,
-                currency = item.currency,
-                date = item.date,
-            )
-        }
-
-        // Fetch the current items from the Room database
-        val currentItems = tipTrackerDao.getItems(items[0].familyId).first()
-
-        // Determine the items to be inserted or updated
-        val itemsToUpdate = tipEntityList.filter { newItem ->
-            currentItems.none { currentItem -> newItem.id == currentItem.id && newItem == currentItem }
-        }
-
-        // Determine the items to be deleted
-        val itemsToDelete = currentItems.filter { currentItem ->
-            tipEntityList.none { newItem -> newItem.id == currentItem.id }
-        }
-
-        // Update the Room database with the new or changed items
-        tipTrackerDao.updateItems(itemsToUpdate)
-
-        // Delete the items that no longer exist in Firestore
-        tipTrackerDao.deleteItems(itemsToDelete.map { it.id })
-    }
-
-    suspend fun deleteFamilyTipItems(familyId: String) {
-        val currentFamilyItems = tipTrackerDao.getItems(familyId).firstOrNull()
-        if (currentFamilyItems != null) {
-            tipTrackerDao.deleteItems(currentFamilyItems.map { it.id })
-        }
-    }
-
-    // -------------------------------------------------------------- GUIDES
-    suspend fun upsertGuides(items: List<Guide>) {
-        Log.d(TAG, "upsertGuides incomingCount=${items.size}")
-        if (items.isEmpty()) {
-            Log.d(TAG, "upsertGuides skipped because incoming list is empty")
-            return
-        }
-
-        val entityList = items.map { item ->
-            GuideEntity(
-                id = item.id ?: "",
-                familyId = item.familyId,
-                itemName = item.itemName,
-                lastUpdated = item.lastUpdated,
-                description = item.description,
-                visibility = item.visibility,
-                ownerUid = item.ownerUid,
-                contentVersion = item.contentVersion,
-                started = item.started,
-                sections = item.sections,
-                resume = item.resume,
-            )
-        }
-
-        val currentItems = guidesDao.getItems(items.first().familyId).first()
-        val itemsToUpdate = entityList.filter { newItem ->
-            currentItems.none { currentItem -> newItem.id == currentItem.id && newItem == currentItem }
-        }
-
-        Log.d(TAG, "upsertGuides itemsToUpdate=${itemsToUpdate.size}")
-        guidesDao.updateItems(itemsToUpdate)
-    }
-
-    suspend fun updateGuides(items: List<Guide>) {
-        Log.d(TAG, "updateGuides incomingCount=${items.size}")
-        if (items.isEmpty()) {
-            Log.d(TAG, "updateGuides skipped because incoming list is empty")
-            return
-        }
-
-        val entityList = items.map { item ->
-            GuideEntity(
-                id = item.id ?: "",
-                familyId = item.familyId,
-                itemName = item.itemName,
-                lastUpdated = item.lastUpdated,
-                description = item.description,
-                visibility = item.visibility,
-                ownerUid = item.ownerUid,
-                contentVersion = item.contentVersion,
-                started = item.started,
-                sections = item.sections,
-                resume = item.resume,
-            )
-        }
-
-        val currentItems = guidesDao.getItems(items.first().familyId).first()
-        Log.d(TAG, "updateGuides currentRoomCount=${currentItems.size}")
-
-        val itemsToUpdate = entityList.filter { newItem ->
-            currentItems.none { currentItem -> newItem.id == currentItem.id && newItem == currentItem }
-        }
-
-        val itemsToDelete = currentItems.filter { currentItem ->
-            entityList.none { newItem -> newItem.id == currentItem.id }
-        }
-
-        Log.d(
-            TAG,
-            "updateGuides itemsToUpdate=${itemsToUpdate.size} itemsToDelete=${itemsToDelete.size}",
-        )
-        val deletedGuideIds = itemsToDelete.map { it.id }
-        guidesDao.updateItems(itemsToUpdate)
-        if (deletedGuideIds.isNotEmpty()) {
-            guideProgressDao.deleteByGuideIds(items.first().familyId, deletedGuideIds)
-            guidesDao.deleteItems(deletedGuideIds)
-        }
-    }
-
-    suspend fun deleteFamilyGuides(familyId: String) {
-        val currentFamilyItems = guidesDao.getItems(familyId).first()
-        if (currentFamilyItems.isNotEmpty()) {
-            Log.d(TAG, "deleteFamilyGuides familyId=$familyId count=${currentFamilyItems.size}")
-            val deletedGuideIds = currentFamilyItems.map { it.id }
-            guideProgressDao.deleteByGuideIds(familyId, deletedGuideIds)
-            guidesDao.deleteItems(deletedGuideIds)
-        }
-    }
-
-    // -------------------------------------------------------------- USER LISTS
-    suspend fun upsertUserLists(items: List<UserList>) {
-        Log.d(TAG, "upsertUserLists incomingCount=${items.size}")
-        if (items.isEmpty()) return
-        val entityList = items.map { it.toEntity() }
-        val currentItems = userListsDao.getItems(items.first().familyId).first()
-        val itemsToUpdate = entityList.filter { newItem ->
-            currentItems.none { cur -> newItem.id == cur.id && newItem == cur }
-        }
-        Log.d(TAG, "upsertUserLists itemsToUpdate=${itemsToUpdate.size}")
-        userListsDao.updateItems(itemsToUpdate)
-    }
-
-    suspend fun updateUserLists(items: List<UserList>) {
-        Log.d(TAG, "updateUserLists incomingCount=${items.size}")
-        if (items.isEmpty()) return
-        val entityList = items.map { it.toEntity() }
-        val currentItems = userListsDao.getItems(items.first().familyId).first()
-        Log.d(TAG, "updateUserLists currentRoomCount=${currentItems.size}")
-        val itemsToUpdate = entityList.filter { newItem ->
-            currentItems.none { cur -> newItem.id == cur.id && newItem == cur }
-        }
-        val itemsToDelete = currentItems.filter { cur ->
-            entityList.none { newItem -> newItem.id == cur.id }
-        }
-        Log.d(TAG, "updateUserLists toUpdate=${itemsToUpdate.size} toDelete=${itemsToDelete.size}")
-        userListsDao.updateItems(itemsToUpdate)
-        if (itemsToDelete.isNotEmpty()) {
-            val deletedIds = itemsToDelete.map { it.id }
-            routineListsDao.deleteByListIds(deletedIds)
-            userListsDao.deleteItems(deletedIds)
-        }
-    }
-
-    fun deleteFamilyUserLists(familyId: String) {
-        Log.d(TAG, "deleteFamilyUserLists familyId=$familyId")
-        routineListsDao.deleteFamilyItems(familyId)
-        userListsDao.deleteFamilyItems(familyId)
-    }
-
-    private fun UserList.toEntity() = UserListEntity(
-        id = id ?: "",
-        familyId = familyId,
-        itemName = itemName,
-        lastUpdated = lastUpdated,
-        dateCreated = dateCreated,
-        type = type,
-        visibility = visibility,
-        ownerUid = ownerUid,
-    )
-
-    // -------------------------------------------------------------- LIST ENTRIES
-    suspend fun updateRoutineListEntries(
-        items: List<RoutineListEntry>,
-        byteArrays: Map<String, ByteArray> = emptyMap(),
-    ) {
-        Log.d(TAG, "updateRoutineListEntries incomingCount=${items.size}")
-        if (items.isEmpty()) return
-        val currentItems = routineListsDao.getItems(items.first().familyId).first()
-        Log.d(TAG, "updateRoutineListEntries currentRoomCount=${currentItems.size}")
-
-        var entityList = items.map { it.toEntity() }
-
-        // Preserve existing imageData for entries not being re-downloaded
-        entityList = entityList.map { entity ->
-            val existingImage = currentItems.find { it.id == entity.id }?.imageData
-            val newImage = byteArrays[entity.id]
-            entity.copy(imageData = newImage ?: existingImage)
-        }
-
-        val itemsToUpdate = entityList.filter { newItem ->
-            currentItems.none { cur -> newItem.id == cur.id && newItem == cur }
-        }
-        val itemsToDelete = currentItems.filter { cur ->
-            entityList.none { newItem -> newItem.id == cur.id }
-        }
-        Log.d(TAG, "updateRoutineListEntries toUpdate=${itemsToUpdate.size} toDelete=${itemsToDelete.size}")
-        routineListsDao.updateItems(itemsToUpdate)
-        if (itemsToDelete.isNotEmpty()) {
-            routineListsDao.deleteItems(itemsToDelete.map { it.id })
-        }
-    }
-
-    suspend fun getRoutineEntryIdsWithImages(familyId: String): List<String> =
-        routineListsDao.getEntryIdsWithImages(familyId)
-
-    fun deleteFamilyRoutineListEntries(familyId: String) {
-        Log.d(TAG, "deleteFamilyRoutineListEntries familyId=$familyId")
-        routineListsDao.deleteFamilyItems(familyId)
-    }
-
-    private fun RoutineListEntry.toEntity() = RoutineListEntryEntity(
-        id = id ?: "",
-        familyId = familyId,
-        listId = listId,
-        itemName = itemName,
-        lastUpdated = lastUpdated,
-        dateCreated = dateCreated,
-        nextDate = nextDate,
-        lastCompletedAt = lastCompletedAt,
-        completionCount = completionCount,
-        recurrenceUnit = recurrenceUnit,
-        interval = interval,
-        weekdays = weekdays,
-        imageData = null,
-    )
-
-    suspend fun upsertGuideProgress(progress: GuideProgressState) {
-        val existing = guideProgressDao.getItemById(progress.id)
-        val entity = GuideProgressEntity(
-            id = progress.id,
-            familyId = progress.familyId,
-            uid = progress.uid,
-            guideId = progress.guideId,
-            contentVersion = progress.contentVersion,
-            started = progress.started,
-            completedPointerKeys = progress.completedPointerKeys,
-            resume = progress.resume,
-            lastUpdated = progress.lastUpdated,
-            pendingSync = true,
-            localUpdatedAt = progress.localUpdatedAt,
-            lastUploadedAt = existing?.lastUploadedAt,
-        )
-        guideProgressDao.upsertItem(entity)
-    }
-
-    suspend fun updateGuideProgressFromRemote(
-        familyId: String,
-        uid: String,
-        items: List<GuideProgressState>,
-    ) {
-        val entityList = mutableListOf<GuideProgressEntity>()
-        val protectedGuideIds = mutableSetOf<String>()
-        items.forEach { item ->
-            val existing = guideProgressDao.getItemById(item.id)
-            if (existing?.pendingSync == true) {
-                protectedGuideIds += existing.guideId
-                return@forEach
-            }
-            entityList += GuideProgressEntity(
-                id = item.id,
-                familyId = item.familyId,
-                uid = item.uid,
-                guideId = item.guideId,
-                contentVersion = item.contentVersion,
-                started = item.started,
-                completedPointerKeys = item.completedPointerKeys,
-                resume = item.resume,
-                lastUpdated = item.lastUpdated,
-                pendingSync = false,
-                localUpdatedAt = item.localUpdatedAt,
-                lastUploadedAt = item.lastUploadedAt ?: item.lastUpdated,
-            )
-        }
-
-        if (entityList.isNotEmpty()) {
-            guideProgressDao.upsertItems(entityList)
-        }
-
-        val localPendingGuideIds = guideProgressDao.getPendingItems(familyId, uid)
-            .map { it.guideId }
-            .toSet()
-        val keepGuideIds = items.map { it.guideId }.toSet() + localPendingGuideIds + protectedGuideIds
-
-        if (keepGuideIds.isEmpty()) {
-            guideProgressDao.deleteFamilyUserItems(familyId, uid)
-            return
-        }
-        guideProgressDao.deleteMissingGuides(familyId, uid, keepGuideIds.toList())
-    }
-
-    suspend fun getPendingGuideProgresses(
-        familyId: String,
-        uid: String,
-        guideId: String? = null,
-    ): List<GuideProgressState> {
-        val allPending = guideProgressDao.getPendingItems(familyId, uid)
-        val filtered = if (guideId.isNullOrBlank()) {
-            allPending
-        } else {
-            allPending.filter { it.guideId == guideId }
-        }
-        return filtered.map { entity ->
-            GuideProgressState(
-                id = entity.id,
-                familyId = entity.familyId,
-                uid = entity.uid,
-                guideId = entity.guideId,
-                contentVersion = entity.contentVersion,
-                started = entity.started,
-                completedPointerKeys = entity.completedPointerKeys,
-                resume = entity.resume,
-                lastUpdated = entity.lastUpdated,
-                pendingSync = entity.pendingSync,
-                localUpdatedAt = entity.localUpdatedAt,
-                lastUploadedAt = entity.lastUploadedAt,
-            )
-        }
-    }
-
-    suspend fun markGuideProgressSynced(
-        progressId: String,
-        uploadedAt: Date,
-    ) {
-        val existing = guideProgressDao.getItemById(progressId) ?: return
-        guideProgressDao.upsertItem(
-            existing.copy(
-                pendingSync = false,
-                lastUploadedAt = uploadedAt,
-            ),
-        )
-    }
-
-    private fun applyProgressToGuideEntity(
-        guide: GuideEntity,
-        progress: GuideProgressEntity?,
-    ): GuideEntity {
-        if (progress == null) return guide
-        if (progress.contentVersion != guide.contentVersion) {
-            return guide.copy(
-                started = false,
-                resume = null,
-                sections = GuideProgressSnapshot.applyCompletedPointerKeys(
-                    sections = guide.sections,
-                    completedPointerKeys = emptySet(),
-                ),
-            )
-        }
-        val mergedLastUpdated = if (progress.lastUpdated.after(guide.lastUpdated)) {
-            progress.lastUpdated
-        } else {
-            guide.lastUpdated
-        }
-        return guide.copy(
-            started = progress.started,
-            sections = GuideProgressSnapshot.applyCompletedPointerKeys(
-                sections = guide.sections,
-                completedPointerKeys = progress.completedPointerKeys.toSet(),
-            ),
-            resume = progress.resume,
-            lastUpdated = mergedLastUpdated,
-        )
-    }
-
-    // -------------------------------------------------------------- IMAGES
-    fun getImageByteArray(imageType: ImageType): Flow<ByteArray?> {
-        Log.d(TAG, "getImageByteArray imageType: $imageType")
-        return when (imageType) {
-            is ImageType.ProfileImage -> userInformationDao.getImageByteArray(imageType.uid)
-
-            is ImageType.FamilyImage -> familyInformationDao.getImageByteArray(imageType.familyId)
-
-            is ImageType.RecipeImage -> recipesDao.getImageByteArray(
-                imageType.familyId,
-                imageType.recipeId,
-            )
-
-            is ImageType.GalleryMedia -> flowOf(null)
-
-            is ImageType.RoutineListEntryImage -> routineListsDao.getImageByteArray(imageType.entryId)
-        }
-    }
-
-    // --- Download helper functions ---
-    fun getMediaFilesForDownload(
-        mediaIds: List<String>,
-        familyId: String,
-    ): List<Pair<File, GalleryMedia?>>? {
-        try {
-            val pairs: MutableList<Pair<File, GalleryMedia?>> = mutableListOf()
-            val mediaItems = galleryMediaDao.getItemsByIds(familyId,mediaIds)
-            mediaItems?.forEach { mediaItem ->
-                mediaItem.mediaUri?.let { path ->
-                    val file = File(path)
-                    if (file.exists()) {
-                        Log.d("LocalDataSource", "Found media file at: $path")
-                        val domainMediaItem = when (mediaItem.mediaType) {
-                            MediaType.IMAGE -> {
-                                GalleryImage(
-                                    id = mediaItem.id,
-                                    familyId = mediaItem.familyId,
-                                    itemName = mediaItem.itemName,
-                                    lastUpdated = mediaItem.lastUpdated,
-                                    albumId = mediaItem.albumId,
-                                    dateCreated = mediaItem.dateCreated,
-                                )
-                            }
-
-                            MediaType.VIDEO -> {
-                                GalleryVideo(
-                                    id = mediaItem.id,
-                                    familyId = mediaItem.familyId,
-                                    itemName = mediaItem.itemName,
-                                    lastUpdated = mediaItem.lastUpdated,
-                                    albumId = mediaItem.albumId,
-                                    dateCreated = mediaItem.dateCreated,
-                                    duration = mediaItem.videoDuration,
-                                )
-                            }
-                        }
-                        pairs.add(Pair(file, domainMediaItem))
-                    } else {
-                        Log.w("LocalDataSource", "Media file path doesn't exist: $path")
-                    }
-                }
-            }
-            return pairs
-        } catch (e: Exception) {
-            Log.e("LocalDataSource", "Error getting media file for download: ${e.message}", e)
-            return null
-        }
-    }
-
-    fun copyMediaToGalleryFolder(
-        mediaFile: File,
-        mediaItem: GalleryMedia,
-    ): ResultListener {
-        val resolver = context.contentResolver
-        var mediaStoreUri: Uri? = null
-
-        return try {
-            val (collectionUri, relativePath, mimeType) = when (mediaItem) {
-                is GalleryImage -> Triple(
-                    MediaStore.Images.Media.getContentUri(MediaStore.VOLUME_EXTERNAL_PRIMARY),
-                    "Pictures/LifeTogether",
-                    "image/jpeg"
-                )
-                is GalleryVideo -> Triple(
-                    MediaStore.Video.Media.getContentUri(MediaStore.VOLUME_EXTERNAL_PRIMARY),
-                    "Movies/LifeTogether",
-                    "video/mp4"
-                )
-            }
-
-            // 2. Prepare Metadata
-            val contentValues = ContentValues().apply {
-                put(MediaStore.MediaColumns.DISPLAY_NAME, mediaItem.itemName)
-                put(MediaStore.MediaColumns.MIME_TYPE, mimeType)
-                put(MediaStore.MediaColumns.RELATIVE_PATH, relativePath)
-                put(MediaStore.MediaColumns.IS_PENDING, 1)
-
-                mediaItem.dateCreated?.time?.let {
-                    put(MediaStore.MediaColumns.DATE_TAKEN, it)
-                }
-
-                if (mediaItem is GalleryVideo) {
-                    mediaItem.duration?.let { put(MediaStore.Video.Media.DURATION, it) }
-                }
-            }
-
-            // 3. Insert Empty File (Reserve the spot)
-            mediaStoreUri = resolver.insert(collectionUri, contentValues)
-                ?: return ResultListener.Failure("Failed to create MediaStore entry")
-
-            // 4. Write Data to Stream
-            resolver.openOutputStream(mediaStoreUri)?.use { outputStream ->
-                mediaFile.inputStream().use { inputStream ->
-                    inputStream.copyTo(outputStream)
-                }
-            } ?: return ResultListener.Failure("Failed to open output stream")
-
-            // 5. Finalize (Publish the file)
-            contentValues.clear()
-            contentValues.put(MediaStore.MediaColumns.IS_PENDING, 0)
-            resolver.update(mediaStoreUri, contentValues, null, null)
-
-            Log.d("LocalDataSource", "Media saved successfully: $mediaStoreUri")
-            ResultListener.Success
-
-        } catch (e: Exception) {
-            Log.e("LocalDataSource", "Error saving media: ${e.message}", e)
-
-            // Cleanup: Delete the corrupted/empty file if we crashed midway
-            mediaStoreUri?.let { uri ->
-                try {
-                    resolver.delete(uri, null, null)
-                } catch (_: Exception) {
-                    Log.w("LocalDataSource", "Failed to clean up partial file: $uri")
-                }
-            }
-
-            ResultListener.Failure("Failed to save file: ${e.message}")
-        }
-    }
-}
