+++ app/src/main/java/com/example/lifetogether/domain/usecase/observers/ObserveFamilyInformationUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.lifetogether.domain.usecase.observers
 
-import com.example.lifetogether.data.local.LocalDataSource
+import com.example.lifetogether.data.local.source.UserLocalDataSource
 import com.example.lifetogether.data.remote.FirestoreDataSource
 import com.example.lifetogether.domain.repository.StorageRepository
 import com.example.lifetogether.domain.listener.ByteArrayResultListener
@@ -13,7 +13,7 @@ import javax.inject.Inject
 class ObserveFamilyInformationUseCase @Inject constructor(
     private val firestoreDataSource: FirestoreDataSource,
     private val storageRepository: StorageRepository,
-    private val localDataSource: LocalDataSource,
+    private val userLocalDataSource: UserLocalDataSource,
 ) {
     fun start(
         scope: CoroutineScope,
@@ -29,7 +29,7 @@ class ObserveFamilyInformationUseCase @Inject constructor(
                         runCatching {
                             // Check if family already has an image to avoid re-downloading
                             val hasExistingImage = result.familyInformation.familyId?.let { familyIdValue ->
-                                localDataSource.familyHasImage(familyIdValue)
+                                userLocalDataSource.familyHasImage(familyIdValue)
                             } ?: false
 
                             if (!hasExistingImage) {
@@ -43,7 +43,7 @@ class ObserveFamilyInformationUseCase @Inject constructor(
 
                                 when (byteArrayResult) {
                                     is ByteArrayResultListener.Success -> {
-                                        localDataSource.updateFamilyInformation(
+                                        userLocalDataSource.updateFamilyInformation(
                                             result.familyInformation,
                                             byteArrayResult.byteArray,
                                         )
@@ -52,12 +52,12 @@ class ObserveFamilyInformationUseCase @Inject constructor(
                                     is ByteArrayResultListener.Failure -> {
                                         println("ByteArrayResultListener failure: ${byteArrayResult.message}")
                                         // Update without image on failure
-                                        localDataSource.updateFamilyInformation(result.familyInformation)
+                                        userLocalDataSource.updateFamilyInformation(result.familyInformation)
                                     }
 
                                     null -> {
                                         // No image URL provided, update without image
-                                        localDataSource.updateFamilyInformation(result.familyInformation)
+                                        userLocalDataSource.updateFamilyInformation(result.familyInformation)
                                     }
                                 }
                             } else {
