+++ app/src/main/java/com/example/lifetogether/domain/usecase/observers/ObserveTipTrackerUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.lifetogether.domain.usecase.observers
 
-import com.example.lifetogether.data.local.LocalDataSource
+import com.example.lifetogether.data.local.source.TipTrackerLocalDataSource
 import com.example.lifetogether.data.remote.FirestoreDataSource
 import com.example.lifetogether.domain.listener.ListItemsResultListener
 import kotlinx.coroutines.CompletableDeferred
@@ -10,7 +10,7 @@ import javax.inject.Inject
 
 class ObserveTipTrackerUseCase @Inject constructor(
     private val firestoreDataSource: FirestoreDataSource,
-    private val localDataSource: LocalDataSource,
+    private val tipTrackerLocalDataSource: TipTrackerLocalDataSource,
 ) {
     fun start(
         scope: CoroutineScope,
@@ -26,9 +26,9 @@ class ObserveTipTrackerUseCase @Inject constructor(
                         runCatching {
                             if (result.listItems.isEmpty()) {
                                 println("tipTrackerSnapshotListener().collect result: is empty")
-                                localDataSource.deleteFamilyTipItems(familyId)
+                                tipTrackerLocalDataSource.deleteFamilyTipItems(familyId)
                             } else {
-                                localDataSource.updateTipTracker(result.listItems)
+                                tipTrackerLocalDataSource.updateTipTracker(result.listItems)
                             }
                         }.onSuccess {
                             firstSuccess.completeFirstSuccessIfNeeded()
