+++ app/src/main/java/com/example/lifetogether/domain/usecase/observers/ObserveUserInformationUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.lifetogether.domain.usecase.observers
 
-import com.example.lifetogether.data.local.LocalDataSource
+import com.example.lifetogether.data.local.source.UserLocalDataSource
 import com.example.lifetogether.data.remote.FirestoreDataSource
 import com.example.lifetogether.domain.repository.StorageRepository
 import com.example.lifetogether.domain.listener.AuthResultListener
@@ -13,7 +13,7 @@ import javax.inject.Inject
 class ObserveUserInformationUseCase @Inject constructor(
     private val firestoreDataSource: FirestoreDataSource,
     private val storageRepository: StorageRepository,
-    private val localDataSource: LocalDataSource,
+    private val userLocalDataSource: UserLocalDataSource,
 ) {
     fun start(
         scope: CoroutineScope,
@@ -29,7 +29,7 @@ class ObserveUserInformationUseCase @Inject constructor(
                         runCatching {
                             // Check if user already has a profile image to avoid re-downloading
                             val hasExistingImage = result.userInformation.uid?.let { uidValue ->
-                                localDataSource.userHasProfileImage(uidValue)
+                                userLocalDataSource.userHasProfileImage(uidValue)
                             } ?: false
 
                             if (!hasExistingImage) {
@@ -41,7 +41,7 @@ class ObserveUserInformationUseCase @Inject constructor(
 
                                 when (byteArrayResult) {
                                     is ByteArrayResultListener.Success -> {
-                                        localDataSource.updateUserInformation(
+                                        userLocalDataSource.updateUserInformation(
                                             result.userInformation,
                                             byteArrayResult.byteArray,
                                         )
@@ -50,12 +50,12 @@ class ObserveUserInformationUseCase @Inject constructor(
                                     is ByteArrayResultListener.Failure -> {
                                         println("ByteArrayResultListener failure: ${byteArrayResult.message}")
                                         // Update without image on failure
-                                        localDataSource.updateUserInformation(result.userInformation)
+                                        userLocalDataSource.updateUserInformation(result.userInformation)
                                     }
 
                                     null -> {
                                         // No image URL provided, update without image
-                                        localDataSource.updateUserInformation(result.userInformation)
+                                        userLocalDataSource.updateUserInformation(result.userInformation)
                                     }
                                 }
                             } else {
